requirements.txt
fastmcp
playwright
python-frontmatter
aiofiles 